more moving grapher to new backend

+++ trunk/experimental/contentprovider/src/net/redgeek/android/eventgrapher/primitives/CategoryDatapointCache.java
@@ -24,15 +24,15 @@ import java.util.TreeMap;
 public class CategoryDatapointCache {
   private long mCatId;
   private boolean mValid;
-  private TreeMap<Long, Datapoint> mCache;
-  private long mStart;
-  private long mEnd;
+  private TreeMap<Integer, Datapoint> mCache;
+  private int mStart;
+  private int mEnd;
   private int mHistory;
 
   public CategoryDatapointCache(long catId, int history) {
     mCatId = catId;
     mValid = false;
-    mCache = new TreeMap<Long, Datapoint>();
+    mCache = new TreeMap<Integer, Datapoint>();
     resetRangeMarkers();
     mHistory = history;
   }
@@ -50,20 +50,20 @@ public class CategoryDatapointCache {
     return mCatId;
   }
 
-  public long getStart() {
+  public int getStart() {
     return mStart;
   }
 
-  public void updateStart(long start) {
+  public void updateStart(int start) {
     if (start < mStart)
       mStart = start;
   }
 
-  public long getEnd() {
+  public int getEnd() {
     return mEnd;
   }
 
-  public void updateEnd(long end) {
+  public void updateEnd(int end) {
     if (end > mEnd)
       mEnd = end;
   }
@@ -91,15 +91,15 @@ public class CategoryDatapointCache {
     return mCache.put(d.mTsStart, d);
   }
 
-  public ArrayList<Datapoint> getDataInRange(long msStart, long msEnd) {
+  public ArrayList<Datapoint> getDataInRange(int msStart, int msEnd) {
     ArrayList<Datapoint> range = new ArrayList<Datapoint>();
-    SortedMap<Long, Datapoint> map;
+    SortedMap<Integer, Datapoint> map;
 
     if (msStart > msEnd)
       return range;
     
     try {
-      map = mCache.subMap(Long.valueOf(msStart), Long.valueOf(msEnd + 1));
+      map = mCache.subMap(Integer.valueOf(msStart), Integer.valueOf(msEnd + 1));
     } catch (NullPointerException e) {
       return range;
     }
@@ -114,20 +114,20 @@ public class CategoryDatapointCache {
     return range;
   }
 
-  public ArrayList<Datapoint> getDataBefore(int number, long ms) {
+  public ArrayList<Datapoint> getDataBefore(int number, int ms) {
     ArrayList<Datapoint> pre = new ArrayList<Datapoint>();
-    SortedMap<Long, Datapoint> range;
-    SortedMap<Long, Datapoint> reverse;
+    SortedMap<Integer, Datapoint> range;
+    SortedMap<Integer, Datapoint> reverse;
     
     try {
-      range = mCache.headMap(Long.valueOf(ms));
+      range = mCache.headMap(Integer.valueOf(ms));
     } catch (NullPointerException e) {
       return pre;
     } catch (IllegalArgumentException e) {
       return pre;
     }
     
-    reverse = new TreeMap<Long, Datapoint>(java.util.Collections.reverseOrder());
+    reverse = new TreeMap<Integer, Datapoint>(java.util.Collections.reverseOrder());
     reverse.putAll(range);
 
     Iterator<Datapoint> iterator = reverse.values().iterator();
@@ -142,12 +142,12 @@ public class CategoryDatapointCache {
     return pre;
   }
 
-  public ArrayList<Datapoint> getDataAfter(int number, long ms) {
+  public ArrayList<Datapoint> getDataAfter(int number, int ms) {
     ArrayList<Datapoint> post = new ArrayList<Datapoint>();
-    SortedMap<Long, Datapoint> range;
+    SortedMap<Integer, Datapoint> range;
 
     try {
-      range = mCache.tailMap(Long.valueOf(ms) + 1);
+      range = mCache.tailMap(Integer.valueOf(ms) + 1);
     } catch (NullPointerException e) {
       return post;
     } catch (IllegalArgumentException e) {
@@ -168,7 +168,7 @@ public class CategoryDatapointCache {
 
   public ArrayList<Datapoint> getLast(int number) {
     ArrayList<Datapoint> last = new ArrayList<Datapoint>();
-    SortedMap<Long, Datapoint> reverse = new TreeMap<Long, Datapoint>(
+    SortedMap<Integer, Datapoint> reverse = new TreeMap<Integer, Datapoint>(
         java.util.Collections.reverseOrder());
     reverse.putAll(mCache);
 
@@ -186,7 +186,7 @@ public class CategoryDatapointCache {
 
   private void resetRangeMarkers() {
     mValid = false;
-    mStart = Long.MAX_VALUE;
-    mEnd = Long.MIN_VALUE;
+    mStart = Integer.MAX_VALUE;
+    mEnd = Integer.MIN_VALUE;
   }
 }
